A last bit of serialization performance on the deserialization side.
There's no point converting a Class<?> to a char and then doing (effectively)
an if-else on the char when we can do it on the Class<?> directly. With this
change, the profile for deserialization is quite reasonable: mostly object
creation and ploughing through the bytes. We'll benefit from future Arrays
work, but that's for another day.

Bug: http://code.google.com/p/android/issues/detail?id=13138
Change-Id: I67aa237c4176e443f7ecdc12f0534cc2f3df5076